+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,8 @@
 # UNRELEASED
 
+  * [changed] `Config` now supports multiple default config blocks per environnment
+  * [changed] `Config#defaults` now supports passing multiple environments
+
 # 1.0
 
   * Hello, Web